Changed ulong params for RenderSystem::setStencilBufferParams to uint32.

+++ trunk/RenderSystems/Direct3D7/include/OgreD3D7RenderSystem.h
@@ -364,7 +364,7 @@ namespace Ogre {
           RenderSystem
          */
         void setStencilBufferParams(CompareFunction func = CMPF_ALWAYS_PASS, 
-            ulong refValue = 0, ulong mask = 0xFFFFFFFF, 
+            uint32 refValue = 0, uint32 mask = 0xFFFFFFFF, 
             StencilOperation stencilFailOp = SOP_KEEP, 
             StencilOperation depthFailOp = SOP_KEEP,
             StencilOperation passOp = SOP_KEEP,
