+++ trunk/RenderSystems/Direct3D7/src/OgreD3D7RenderSystem.cpp
@@ -2395,8 +2395,8 @@ namespace Ogre {
 
     }
     //---------------------------------------------------------------------
-    void D3DRenderSystem::setStencilBufferParams(CompareFunction func, ulong refValue, 
-        ulong mask, StencilOperation stencilFailOp, 
+    void D3DRenderSystem::setStencilBufferParams(CompareFunction func, 
+        uint32 refValue, uint32 mask, StencilOperation stencilFailOp, 
         StencilOperation depthFailOp, StencilOperation passOp, 
         bool twoSidedOperation)
     {
